replaced PdfPoint with Vector2

+++ src/IxMilia.Pdf/Extensions/WriterExtensions.cs
@@ -14,7 +14,7 @@ namespace IxMilia.Pdf.Extensions
             return value.ToString(CultureInfo.InvariantCulture);
         }
 
-        public static string AsFixed(this double value)
+        public static string AsFixed(this float value)
         {
             var str = value.ToString("f2", CultureInfo.InvariantCulture);
             // special case for really small negative values that round to 0
